Rename DrawerNew to Drawer as its the only drawer component now
